Converted to using Maven.
It has involved quite a few renames and Eclipse throw-ups, but it seems to have worked. SpoutAPI is included in the POM but it is not used yet, but I am still working on a native port of Shortify to Spout.

This marks the 1.4.0 release. It has been quite a while to get this far.